LRDOCS-8814 Rename
