+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/creating-a-custom-osgi-service/resources/liferay-s1j6.zip/s1j6-able-impl/src/main/java/com/acme/s1j6/able/internal/S1J6AbleImpl.java
@@ -1,15 +1,15 @@
-package com.acme.s1j6.internal;
+package com.acme.s1j6.able.internal;
 
 import com.acme.s1j6.S1J6;
 
 import org.osgi.service.component.annotations.Component;
 
 @Component(service = S1J6.class)
-public class S1J6Impl implements S1J6 {
+public class S1J6AbleImpl implements S1J6 {
 
 	@Override
 	public String doSomething() {
-		return "S1J6";
+		return "This is the S1J6 Able implementation.";
 	}
 
 }
