+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/creating-a-custom-osgi-service/resources/liferay-s1j6.zip/s1j6-baker-impl/src/main/java/com/acme/s1j6/baker/internal/S1J6BakerImpl.java
@@ -1,4 +1,4 @@
-package com.acme.s1j6.internal.custom;
+package com.acme.s1j6.baker.internal;
 
 import com.acme.s1j6.S1J6;
 
@@ -6,23 +6,18 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 @Component(property = "service.ranking:Integer=100", service = S1J6.class)
-public class S1J6CustomImpl implements S1J6 {
+public class S1J6BakerImpl implements S1J6 {
 
 	@Override
 	public String doSomething() {
-		StringBuilder sb = new StringBuilder();
-
-		sb.append(
-			"S1J6CustomImpl, which delegates to " +
-				_defaultService.doSomething());
-
-		return sb.toString();
+		return _s1j6.doSomething() +
+			"<br />This is the S1J6 Baker implementation.";
 	}
 
 	@Reference(
 		target = "(component.name=com.acme.s1j6.internal.S1J6Impl)",
 		unbind = "-"
 	)
-	private S1J6 _defaultService;
+	private S1J6 _s1j6;
 
 }
