--- docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/creating-a-custom-osgi-service/resources/liferay-s1j6.zip/s1j6-custom-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Acme S1J6 Custom Implementation
-Bundle-SymbolicName: com.acme.s1j6.custom.impl
-Bundle-Version: 1.0.0
