+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/creating-a-custom-osgi-service/resources/liferay-s1j6.zip/s1j6-web/src/main/java/com/acme/s1j6/web/internal/portlet/S1J6Portlet.java
@@ -7,7 +7,6 @@ import java.io.PrintWriter;
 
 import javax.portlet.GenericPortlet;
 import javax.portlet.Portlet;
-import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
@@ -24,13 +23,13 @@ import org.osgi.service.component.annotations.Reference;
 public class S1J6Portlet extends GenericPortlet {
 
 	@Override
-	protected void doView(RenderRequest request, RenderResponse response)
-		throws IOException, PortletException {
+	protected void doView(
+			RenderRequest renderRequest, RenderResponse renderResponse)
+		throws IOException {
 
-		PrintWriter printWriter = response.getWriter();
+		PrintWriter printWriter = renderResponse.getWriter();
 
-		printWriter.println("I'm calling a service ...<br>");
-
+		printWriter.println("This is the portlet.<br />");
 		printWriter.println(_s1J6.doSomething());
 	}
